refactor(swift): split ContentView.swift into focused files
Extract EventRow, ToolCallBadge, CollapsedToolSummary, ApprovalView,
MicButton, ThinkingRow, ChatTimelineView, InputBarView, ToolPhase
into separate files.
