+++ DragonglassApp/Dragonglass/ContentView.swift
@@ -24,10 +24,26 @@ struct ContentView: View {
             if backend.phase != .ready {
                 statusView
             } else {
-                chatView
+                ChatTimelineView(
+                    includeToolCallsInSelection: includeToolCallsInSelection,
+                    isAtBottom: $isAtBottom,
+                    onEventsChanged: handleEventsChanged,
+                    onResend: { msg in inputText = msg }
+                )
+                .environmentObject(client)
+                .environmentObject(backend)
             }
 
-            inputArea
+            InputBarView(
+                inputText: $inputText,
+                onSend: sendMessage,
+                onSTT: sendSTT
+            )
+            .environmentObject(client)
+            .environmentObject(sttManager)
+            .environmentObject(backend)
+            .onAppear { installEscapeMonitor() }
+            .onDisappear { removeEscapeMonitor() }
         }
         .frame(width: 400, height: 500)
         .sheet(item: $client.pendingApproval) { req in
@@ -222,180 +238,18 @@ struct ContentView: View {
         }
     }
 
-    private var chatView: some View {
-        ScrollViewReader { proxy in
-            ScrollView {
-                LazyVStack(alignment: .leading, spacing: 10) {
-                    if let warning = backend.obsidianWarning {
-                        HStack(spacing: 6) {
-                            Image(systemName: "exclamationmark.triangle")
-                                .foregroundColor(.orange)
-                                .font(.caption)
-                            Text(warning)
-                                .font(.caption)
-                                .foregroundColor(.secondary)
-                        }
-                        .padding(8)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .background(Color.orange.opacity(0.08))
-                        .cornerRadius(6)
-                    }
-                    ForEach(client.prefixEventIndices, id: \.self) { index in
-                        EventRow(event: client.events[index], detailed: client.detailedToolEvents, includeToolCalls: includeToolCallsInSelection)
-                    }
-
-                    ForEach(client.turns) { turn in
-                        if !turn.isCompleted, case .userMessage(let msg) = client.events[turn.userMessageIndex] {
-                            HStack(alignment: .top, spacing: 4) {
-                                Button(action: {
-                                    inputText = msg
-                                }) {
-                                    Image(systemName: "arrow.counterclockwise")
-                                        .font(.caption)
-                                        .foregroundColor(.secondary)
-                                        .padding(4)
-                                }
-                                .buttonStyle(.plain)
-                                .help("Resend message")
-                                EventRow(event: client.events[turn.userMessageIndex], detailed: client.detailedToolEvents, includeToolCalls: includeToolCallsInSelection)
-                            }
-                        } else {
-                            EventRow(event: client.events[turn.userMessageIndex], detailed: client.detailedToolEvents, includeToolCalls: includeToolCallsInSelection)
-                        }
-
-                        if turn.isCompleted {
-                            CollapsedToolSummary(
-                                turn: turn,
-                                events: client.events,
-                                detailed: client.detailedToolEvents,
-                                onOpenNote: { client.openNote(path: $0) },
-                                includeToolCalls: includeToolCallsInSelection
-                            )
-                            if let doneIdx = turn.doneIndex {
-                                EventRow(event: client.events[doneIdx], detailed: client.detailedToolEvents, includeToolCalls: includeToolCallsInSelection)
-                            }
-                        } else {
-                            if let idx = turn.toolCallIndices.last,
-                               case .mcpTool(let t, let p, let m, let d) = client.events[idx] {
-                                ToolCallBadge(
-                                    tool: t,
-                                    phase: p,
-                                    message: m,
-                                    detail: d,
-                                    detailed: client.detailedToolEvents,
-                                    onOpenNote: { client.openNote(path: $0) },
-                                    selectable: includeToolCallsInSelection
-                                )
-                                    .id(idx)
-                                    .transition(.asymmetric(
-                                        insertion: .move(edge: .bottom).combined(with: .opacity),
-                                        removal: .move(edge: .top).combined(with: .opacity)
-                                    ))
-                            }
-                        }
-
-                        if let aIdx = turn.assistantMessageIndex {
-                            EventRow(event: client.events[aIdx], detailed: client.detailedToolEvents, includeToolCalls: includeToolCallsInSelection)
-                        }
-                    }
-
-                    if client.isThinking {
-                        ThinkingRow()
-                    }
-                }
-                .padding()
-
-                GeometryReader { geo in
-                    Color.clear.preference(
-                        key: BottomVisibilityKey.self,
-                        value: geo.frame(in: .global).minY
-                    )
-                }
-                .frame(height: 1)
-                .id("bottom")
-            }
-            .coordinateSpace(name: "scroll")
-            .backgroundPreferenceValue(BottomVisibilityKey.self) { minY in
-                GeometryReader { scrollGeo in
-                    let scrollMaxY = scrollGeo.frame(in: .global).maxY
-                    Color.clear.onAppear {
-                        isAtBottom = minY <= scrollMaxY
-                    }.onChange(of: minY) { _, y in
-                        isAtBottom = y <= scrollMaxY
-                    }
-                }
-            }
-            .onChange(of: client.events.count) { _, _ in
-                withAnimation(.easeInOut(duration: 0.3)) {}
-                if let last = client.events.indices.last {
-                    if isAtBottom {
-                        proxy.scrollTo("bottom")
-                    }
-
-                    if case .done = client.events[last],
-                       let model = lastModelSent,
-                       !model.isEmpty,
-                       !client.availableModels.contains(model),
-                       !client.extraModels.contains(model),
-                       shouldPersistCustomModel(completedEventIndex: last) {
-                        client.saveModel(model)
-                    }
-                    lastModelSent = nil
-                    lastRequestStartIndex = nil
-                }
-            }
+    private func handleEventsChanged() {
+        guard let last = client.events.indices.last else { return }
+        if case .done = client.events[last],
+           let model = lastModelSent,
+           !model.isEmpty,
+           !client.availableModels.contains(model),
+           !client.extraModels.contains(model),
+           shouldPersistCustomModel(completedEventIndex: last) {
+            client.saveModel(model)
         }
-    }
-
-    private var inputArea: some View {
-        HStack {
-            TextField(sttPendingPrompt, text: $inputText)
-                .textFieldStyle(.plain)
-                .onSubmit(sendMessage)
-                .disabled(client.isThinking)
-
-            MicButton()
-                .environmentObject(sttManager)
-                .environmentObject(client)
-
-            if client.isThinking && client.pendingApproval == nil {
-                Button(action: { client.stopChat() }) {
-                    Image(systemName: "stop.fill")
-                        .font(.body)
-                        .foregroundColor(.red)
-                        .frame(width: 30, height: 30)
-                        .background(Color.red.opacity(0.1))
-                        .cornerRadius(4)
-                }
-                .buttonStyle(.plain)
-            } else if !client.isThinking {
-                Button(action: sendMessage) {
-                    Image(systemName: "paperplane.fill")
-                }
-                .disabled(inputText.isEmpty || backend.phase != .ready)
-            }
-        }
-        .padding()
-        .background(Color(NSColor.controlBackgroundColor))
-        .onChange(of: sttManager.pendingText) { _, text in
-            if let text { inputText = text }
-        }
-        .onChange(of: sttManager.readyToFire) { _, ready in
-            guard ready, let text = sttManager.pendingText, !text.isEmpty else { return }
-            sttManager.clearFireFlag()
-            let toSend = text
-            inputText = ""
-            sendSTT(toSend)
-        }
-        .onAppear { installEscapeMonitor() }
-        .onDisappear { removeEscapeMonitor() }
-    }
-
-    private var sttPendingPrompt: String {
-        if sttManager.isRecording { return "Recording…" }
-        if sttManager.isTranscribing { return "Transcribing…" }
-        if sttManager.pendingText != nil { return "Press Esc to cancel" }
-        return "Ask anything..."
+        lastModelSent = nil
+        lastRequestStartIndex = nil
     }
 
     private func sendSTT(_ text: String) {
@@ -477,363 +331,5 @@ struct ContentView: View {
         }
 
         return hasText
-    }
-}
-
-struct EventRow: View {
-    let event: AgentEvent
-    var detailed: Bool = false
-    var includeToolCalls: Bool = false
-
-    var body: some View {
-        switch event {
-        case .status(let msg):
-            Text(msg)
-                .font(.caption)
-                .foregroundColor(.secondary)
-                .italic()
-                .textSelection(.enabled)
-        case .assistantMessage(let msg):
-            Text(LocalizedStringKey(msg))
-                .textSelection(.enabled)
-        case .mcpTool:
-            EmptyView()
-        case .config:
-            EmptyView()
-        case .configAck:
-            Text("Settings saved")
-                .font(.caption)
-                .foregroundColor(.green)
-                .textSelection(.enabled)
-        case .done:
-            Divider()
-        case .modelsList:
-            EmptyView()
-        case .usage:
-            EmptyView()
-        case .conversationsList:
-            EmptyView()
-        case .conversationLoaded:
-            EmptyView()
-        case .userMessage(let msg):
-            HStack {
-                Spacer()
-                Text(msg)
-                    .padding(8)
-                    .background(Color.accentColor.opacity(0.1))
-                    .cornerRadius(8)
-                    .textSelection(.enabled)
-            }
-        case .approvalRequest(let req):
-            let row = HStack(spacing: 6) {
-                Image(systemName: "pencil.circle")
-                    .foregroundColor(.orange)
-                Text("Pending approval: \(req.description)")
-                    .font(.caption)
-                    .foregroundColor(.orange)
-            }
-            .padding(4)
-            .background(Color.orange.opacity(0.08))
-            .cornerRadius(4)
-            if includeToolCalls {
-                row.textSelection(.enabled)
-            } else {
-                row.textSelection(.disabled)
-            }
-        case .unknown(let type):
-            Text("Unknown event: \(type)")
-                .font(.caption)
-        }
-    }
-}
-
-enum ToolPhase: String {
-    case done = "done"
-    case error = "error"
-    case validationError = "validation_error"
-    case unknown
-
-    init(rawValue: String) {
-        switch rawValue {
-        case "done": self = .done
-        case "error": self = .error
-        case "validation_error": self = .validationError
-        default: self = .unknown
-        }
-    }
-}
-
-struct ToolCallBadge: View {
-    let tool: String
-    let phase: String
-    let message: String
-    let detail: String
-    var detailed: Bool = false
-    var onOpenNote: ((String) -> Void)?
-    var selectable: Bool = false
-    @State private var showingDetail = false
-
-    private var toolPhase: ToolPhase { ToolPhase(rawValue: phase) }
-
-    private var notePath: String? {
-        guard tool == "dragonglass_read_note_with_hash",
-              toolPhase == .done,
-              message.hasPrefix("Reading: ") else { return nil }
-        return String(message.dropFirst("Reading: ".count))
-    }
-
-    private var badgeColor: Color {
-        switch toolPhase {
-        case .error: return .red
-        case .validationError: return .orange
-        default: return .blue
-        }
-    }
-
-    private var isErrorLike: Bool {
-        toolPhase == .error || toolPhase == .validationError
-    }
-
-    private var badgeLabel: String {
-        switch toolPhase {
-        case .error: return "\(tool): error"
-        case .validationError: return "\(tool): validation error"
-        default: return message
-        }
-    }
-
-    var body: some View {
-        let badge = HStack(alignment: .top, spacing: 6) {
-            if toolPhase == .error {
-                Image(systemName: "exclamationmark.circle")
-                    .foregroundColor(.red)
-            } else if toolPhase == .validationError {
-                Image(systemName: "exclamationmark.triangle")
-                    .foregroundColor(.orange)
-            }
-            if detailed {
-                VStack(alignment: .leading, spacing: 2) {
-                    Text("\(tool) [\(phase)]")
-                        .foregroundColor(.secondary)
-                    Text(message + (detail.isEmpty ? "" : " — \(detail)"))
-                }
-            } else {
-                Text(badgeLabel)
-            }
-        }
-        .font(.caption)
-        .padding(4)
-        .background(badgeColor.opacity(0.08))
-        .cornerRadius(4)
-        .onTapGesture {
-            if isErrorLike {
-                showingDetail = true
-            } else if let path = notePath {
-                onOpenNote?(path)
-            }
-        }
-        .popover(isPresented: $showingDetail) {
-            ScrollView {
-                Text(detail.isEmpty ? "No detail available." : detail)
-                    .font(.caption)
-                    .padding()
-                    .frame(maxWidth: 320, alignment: .leading)
-            }
-            .frame(maxHeight: 200)
-        }
-        if selectable {
-            badge.textSelection(.enabled)
-        } else {
-            badge.textSelection(.disabled)
-        }
-    }
-}
-
-struct CollapsedToolSummary: View {
-    let turn: ChatTurn
-    let events: [AgentEvent]
-    var detailed: Bool = false
-    var onOpenNote: ((String) -> Void)?
-    var includeToolCalls: Bool = false
-    @State private var isExpanded = false
-
-    var body: some View {
-        if turn.toolCallIndices.isEmpty { return AnyView(EmptyView()) }
-        return AnyView(
-            DisclosureGroup(isExpanded: $isExpanded) {
-                VStack(alignment: .leading, spacing: 4) {
-                    ForEach(turn.toolCallIndices, id: \.self) { idx in
-                        if case .mcpTool(let t, let p, let m, let d) = events[idx] {
-                            ToolCallBadge(
-                                tool: t,
-                                phase: p,
-                                message: m,
-                                detail: d,
-                                detailed: detailed,
-                                onOpenNote: onOpenNote,
-                                selectable: includeToolCalls
-                            )
-                        }
-                    }
-                }
-                .padding(.top, 2)
-            } label: {
-                let count = turn.toolCallIndices.count
-                Text("\(count) tool call\(count == 1 ? "" : "s")")
-                    .font(.caption)
-                    .foregroundColor(.secondary)
-                    .textSelection(.disabled)
-            }
-        )
-    }
-}
-
-private struct BottomVisibilityKey: PreferenceKey {
-    static let defaultValue: CGFloat = 0
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = nextValue()
-    }
-}
-
-struct DiffView: View {
-    let diff: String
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            ForEach(Array(diff.components(separatedBy: "\n").enumerated()), id: \.offset) { _, line in
-                Text(line.isEmpty ? " " : line)
-                    .font(.system(.caption, design: .monospaced))
-                    .foregroundColor(lineColor(line))
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .background(lineBackground(line))
-            }
-        }
-    }
-
-    private func lineColor(_ line: String) -> Color {
-        if line.hasPrefix("+") { return .green }
-        if line.hasPrefix("-") { return .red }
-        if line.hasPrefix("@") { return .blue }
-        return .primary
-    }
-
-    private func lineBackground(_ line: String) -> Color {
-        if line.hasPrefix("+") { return Color.green.opacity(0.08) }
-        if line.hasPrefix("-") { return Color.red.opacity(0.08) }
-        return .clear
-    }
-}
-
-struct ApprovalView: View {
-    @EnvironmentObject var client: AgentClient
-    let request: ApprovalRequest
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            Text("Approve Edit?")
-                .font(.headline)
-
-            Text(request.description)
-                .font(.subheadline)
-                .foregroundColor(.secondary)
-
-            if !request.diff.isEmpty {
-                ScrollView {
-                    DiffView(diff: request.diff)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                }
-                .frame(maxHeight: 200)
-                .background(Color(NSColor.textBackgroundColor))
-                .cornerRadius(6)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 6)
-                        .stroke(Color.secondary.opacity(0.3), lineWidth: 1)
-                )
-            }
-
-            HStack {
-                Button("Reject") {
-                    client.rejectRequest(request)
-                }
-                .foregroundColor(.red)
-
-                Spacer()
-
-                Button("Approve for Session") {
-                    client.approveRequest(request, forSession: true)
-                }
-
-                Button("Approve") {
-                    client.approveRequest(request, forSession: false)
-                }
-                .buttonStyle(.borderedProminent)
-            }
-        }
-        .padding()
-        .frame(width: 420, height: 340)
-    }
-}
-
-struct MicButton: View {
-    @EnvironmentObject var sttManager: STTManager
-    @EnvironmentObject var client: AgentClient
-    @State private var pulsing = false
-
-    var body: some View {
-        ZStack {
-            if sttManager.isRecording {
-                Circle()
-                    .fill(Color.red.opacity(0.25))
-                    .scaleEffect(pulsing ? 1.5 : 1.0)
-                    .animation(.easeInOut(duration: 0.7).repeatForever(autoreverses: true), value: pulsing)
-                    .frame(width: 30, height: 30)
-                    .onAppear { pulsing = true }
-                    .onDisappear { pulsing = false }
-            }
-            if sttManager.isTranscribing {
-                ProgressView()
-                    .scaleEffect(0.6)
-                    .frame(width: 30, height: 30)
-            } else {
-                Image(systemName: sttManager.isRecording ? "microphone.fill" : "microphone")
-                    .foregroundColor(sttManager.isRecording ? .red : .secondary)
-                    .frame(width: 30, height: 30)
-            }
-        }
-        .onTapGesture {
-            if sttManager.isRecording {
-                sttManager.stopAndTranscribe()
-            } else {
-                sttManager.startRecording()
-            }
-        }
-        .disabled(!sttManager.micPermissionGranted || !sttManager.isModelReady || client.isThinking)
-        .opacity(sttManager.micPermissionGranted && sttManager.isModelReady ? 1.0 : 0.3)
-        .help(
-            !sttManager.micPermissionGranted ? "Microphone permission required" :
-            !sttManager.isModelReady ? "Download a Whisper model in Settings first" :
-            "Hold to dictate"
-        )
-    }
-}
-
-struct ThinkingRow: View {
-    @State private var opMessage: String = "Thinking"
-
-    var body: some View {
-        HStack {
-            if #available(macOS 14.0, *) {
-                Image(systemName: "sparkles")
-                    .symbolEffect(.pulse)
-            } else {
-                Image(systemName: "sparkles")
-            }
-            Text(opMessage)
-                .font(.caption)
-                .foregroundColor(.secondary)
-        }
-        .padding(8)
-        .background(Color.secondary.opacity(0.1))
-        .cornerRadius(8)
     }
 }
